changed percent sus

+++ models/models.go
@@ -2,10 +2,10 @@ package models
 
 type DataRes struct {
 	Metrics struct {
-		NumOfSusWords int `json:"numberOfSusWords"`
-		NumOfSusChars int `json:"numberOfSusCharacters"`
-		TotalChars    int `json:"totalNumberOfCharacters"`
-		PercentSus    int `json:"percentageOfSusToTotal"`
+		NumOfSusWords int     `json:"numberOfSusWords"`
+		NumOfSusChars int     `json:"numberOfSusCharacters"`
+		TotalChars    int     `json:"totalNumberOfCharacters"`
+		PercentSus    float32 `json:"percentageOfSusToTotal"`
 	} `json:"metrics"`
 	Content string `json:"content"`
 }
